OpenUP/requirements.txt
# Core GUI
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1
pyarrow>=14.0.0

# Image Processing
Pillow>=10.1.0
pillow-heif>=0.13.0
imageio>=2.31.0

# Document Parsing
PyMuPDF>=1.23.0
python-docx>=1.1.0
python-pptx>=0.6.23
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Audio/Video
opencv-python>=4.8.0

# 3D Visualization
open3d>=0.18.0
trimesh>=4.0.0
pyvista>=0.43.0

# Geospatial
rasterio>=1.3.9
laspy>=2.5.0
pyproj>=3.6.0
shapely>=2.0.0

# Medical/Scientific
pydicom>=2.4.0
nibabel>=5.1.0
scipy>=1.11.0
h5py>=3.10.0

# Code Highlighting
Pygments>=2.16.0

# Archive Handling
rarfile>=4.1

# Utilities
chardet>=5.2.0
python-magic>=0.4.27
humanize>=4.9.0
watchdog>=3.0.0

# Logging & Monitoring
colorlog>=6.8.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Packaging
pyinstaller>=6.2.0

# Optional (install separately if needed)
# ffmpeg-python>=0.2.0  # For advanced video processing
# av>=11.0.0  # Alternative video processing
